% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_terra.R
\name{check_terra}
\alias{check_terra}
\title{Internal. Angle associated to the spatial gradient}
\usage{
check_terra(r)
}
\arguments{
\item{r}{A raster type object to be checked for compatibility with terra::SpatRaster}
}
\description{
Internal. Angle associated to the spatial gradient
}
\author{
Casey O'Hara
check_terra(r)
}
