% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempTrend.R, R/tempTrend2.R
\name{tempTrend}
\alias{tempTrend}
\alias{tempTrend2}
\title{Long-term local climatic trends}
\usage{
tempTrend(r, th)

tempTrend(clim_r, th)
}
\arguments{
\item{r}{\code{RasterStack} containing a time series of (annual, seasonal, monthly...) values of
the climatic variable for the period of interest.}

\item{th}{\code{Integer} minimum number of observations in the series needed to
calculate the trend at each cell; default 10.}

\item{clim_r}{\code{SpatRaster} containing a time series of values of
the climatic variable for the period of interest.  Note that the calculated
trend will be in units of (climate variable units) per (time interval of
input raster layers).  For example, if the input raster \code{r} contains
annual values of temperature in degrees Celsius, the resulting trend will
be in degrees Celsius per year.}
}
\value{
A \code{RasterStack} containing the cell-specific temporal trends
extracted from simple linear regressions of the climatic variable against time
("slpTrends" in degree Celsius per year), together with their standard
errors ("seTrends") and statistical significance ("sigTrends").

A \code{SpatRaster} containing the cell-specific temporal trends
  extracted from simple linear regressions of the climatic variable against time
  ("slp_trends" in climate variable units per time interval of input raster
  \code{r}), together with their standard
  errors ("se_trends") and statistical significance ("pval_trends").
}
\description{
Function to calculate temporal trend from a raster series
of a climatic variable. This trend is to be used for the calculation of the
gradient-based climate velocity using gVoCC.

Function to calculate temporal trend from a raster series
of a climatic variable. This trend is to be used for the calculation of the
gradient-based climate velocity using gVoCC.
}
\examples{

yrSST <- sumSeries(HSST, p = "1969-01/2009-12", yr0 = "1955-01-01", l = nlayers(HSST),
fun = function(x) colMeans(x, na.rm = TRUE), freqin = "months", freqout = "years")

# Mean annual SST trend (minimum threshold of 10 years of data), with SE and p-values.

tr <- tempTrend(yrSST, th = 10)

plot(tr)


yrSST <- sumSeries(HSST, p = "1969-01/2009-12", yr0 = "1955-01-01", l = nlayers(HSST),
fun = function(x) colMeans(x, na.rm = TRUE), freqin = "months", freqout = "years")

# Mean annual SST trend (minimum threshold of 10 years of data), with SE and p-values.

tr <- tempTrend(yrSST, th = 10)

plot(tr)

}
\seealso{
{\code{\link{spatGrad}}, \code{\link{gVoCC}}}

{\code{\link{spatGrad}}, \code{\link{gVoCC}}}
}
\author{
Jorge Garcia Molinos and Christopher J. Brown
}
